+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Ignoring `opts.first` on `crud.pairs` call
 * External `keydef` compatibility with built-in `merger`
 
+### Added
+
+* Added support for jsonpath indexes for ``crud.select``
+
 ## [0.7.0] - 2021-05-27
 
 ### Fixed